Implement basic bitset parsing
Add rules to parse bitset properties with values given as a sequence
of 0s and 1s, or as a hexedecimal number (e.g. 0xf). Allow up
to 64 bits to be entered in either representation. Convert parsed
string into the internal bitset type of the property value_type
using appropriate semantic action.