Fix bitset conversion functions
To construct a boost dynamic bit set from a ulong, we need to supply
both the number of bits together with the value.

Update Hex conversion function to construct the bit set directly. Count the number of
bits by counting the number of hex digits following the “0x”, converting the
value to ulong using the HexTo convertor as before.