Improve compatibility across Boost versions
Build on platform with older Boost identified an issue with the use
of the "value_name" memebr function of the typed_value class in
program_options. This is only present on Boost 1.50 and higher.

Wrap call to "value_name" in a preprocessor if using Boost's
version macro to only compile it on sufficient Boost.