New Spirit grammar for trivial unit parsing
In properties, may need to parse key/values of the form:

name : real as <quantity> = 1.234 <unit>

The <unit> element will be in the form of a typical scientific unit
expression, e.g "kg m s^-2". Currently, we'd like to parse this
for basic validity (i.e. correct tokens) and reduce to a conversion
factor, ala CLHEP units.

Implement a very basic Spirit parser for this structure. Provide
simple symbol tables for mass, length and time dimensions. Parse
to a vector of symbol,exponent pairs, and reduce this outside
the grammar for simplicity.

Implement basic sanity unit tests.